reusable observations
